requirements.txt

Flask==2.1.1
langchain-community==0.0.7
tavily-python==1.0.0
